move toast to common
