+++ agency-dashboard/src/react-app-env.d.ts
@@ -14,3 +14,5 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // =============================================================================
+
+/// <reference types="react-scripts" />
