+++ common/assets/status-check-white-icon.png
