+++ publisher/src/components/DataUpload/UploadedFiles.tsx
@@ -15,6 +15,7 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // =============================================================================
 
+import { showToast } from "@justice-counts/common/components/Toast";
 import { Permission } from "@justice-counts/common/types";
 import { when } from "mobx";
 import { observer } from "mobx-react-lite";
@@ -25,7 +26,6 @@ import { removeSnakeCase } from "../../utils";
 import downloadIcon from "../assets/download-icon.png";
 import { Badge, BadgeColorMapping, BadgeColors } from "../Badge";
 import { Loader } from "../Loading";
-import { showToast } from "../Toast";
 import {
   ActionButton,
   ActionsContainer,
