+++ publisher/src/components/Reports/CreateReport.tsx
@@ -19,6 +19,7 @@ import {
   palette,
   typography,
 } from "@justice-counts/common/components/GlobalStyles";
+import { showToast } from "@justice-counts/common/components/Toast";
 import {
   CreateReportFormValuesType,
   ReportOverview,
@@ -45,7 +46,6 @@ import {
   TitleWrapper,
 } from "../Forms";
 import { Dropdown } from "../Forms/Dropdown";
-import { showToast } from "../Toast";
 import {
   PublishButton,
   PublishDataWrapper,
